Move traffic-manager Session to a file of its own
Because the state + session file is quite large and hard to navigate.
