+++ cmd/traffic/cmd/manager/internal/state/state.go
@@ -10,7 +10,6 @@ import (
 	"github.com/google/uuid"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/propagation"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
@@ -23,156 +22,6 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/log"
 	"github.com/telepresenceio/telepresence/v2/pkg/tunnel"
 )
-
-type SessionState interface {
-	Cancel()
-	Done() <-chan struct{}
-	LastMarked() time.Time
-	SetLastMarked(lastMarked time.Time)
-	Dials() <-chan *rpc.DialRequest
-	EstablishBidiPipe(context.Context, tunnel.Stream) (tunnel.Endpoint, error)
-	OnConnect(context.Context, tunnel.Stream) (tunnel.Endpoint, error)
-}
-
-type awaitingBidiPipe struct {
-	stream     tunnel.Stream
-	bidiPipeCh chan tunnel.Endpoint
-}
-
-type sessionState struct {
-	sync.Mutex
-	ctx                 context.Context
-	cancel              context.CancelFunc
-	lastMarked          time.Time
-	awaitingBidiPipeMap map[tunnel.ConnID]*awaitingBidiPipe
-	dials               chan *rpc.DialRequest
-}
-
-// EstablishBidiPipe registers the given stream as waiting for a matching stream to arrive in a call
-// to Tunnel, sends a DialRequest to the owner of this sessionState, and then waits. When the call
-// arrives, a BidiPipe connecting the two streams is returned.
-func (ss *sessionState) EstablishBidiPipe(ctx context.Context, stream tunnel.Stream) (tunnel.Endpoint, error) {
-	// Dispatch directly to agent and let the dial happen there
-	bidiPipeCh := make(chan tunnel.Endpoint)
-	id := stream.ID()
-	abp := &awaitingBidiPipe{stream: stream, bidiPipeCh: bidiPipeCh}
-
-	ss.Lock()
-	if ss.awaitingBidiPipeMap == nil {
-		ss.awaitingBidiPipeMap = map[tunnel.ConnID]*awaitingBidiPipe{id: abp}
-	} else {
-		ss.awaitingBidiPipeMap[id] = abp
-	}
-	ss.Unlock()
-
-	// Send dial request to the client/agent
-	dr := &rpc.DialRequest{ConnId: []byte(id),
-		RoundtripLatency: int64(stream.RoundtripLatency()),
-		DialTimeout:      int64(stream.DialTimeout()),
-	}
-	propagator := otel.GetTextMapPropagator()
-	carrier := propagation.MapCarrier{}
-	propagator.Inject(ctx, carrier)
-	dr.TraceContext = carrier
-	select {
-	case <-ss.Done():
-		return nil, status.Error(codes.Canceled, "session cancelled")
-	case ss.dials <- dr:
-	}
-
-	// Wait for the client/agent to connect. Allow extra time for the call
-	ctx, cancel := context.WithTimeout(ctx, stream.DialTimeout()+stream.RoundtripLatency())
-	defer cancel()
-	select {
-	case <-ctx.Done():
-		return nil, status.Error(codes.DeadlineExceeded, "timeout while establishing bidipipe")
-	case <-ss.Done():
-		return nil, status.Error(codes.Canceled, "session cancelled")
-	case bidi := <-bidiPipeCh:
-		return bidi, nil
-	}
-}
-
-// OnConnect checks if a stream is waiting for the given stream to arrive in order to create a BidiPipe.
-// If that's the case, the BidiPipe is created, started, and returned by both this method and the EstablishBidiPipe
-// method that registered the waiting stream. Otherwise, this method returns nil.
-func (ss *sessionState) OnConnect(ctx context.Context, stream tunnel.Stream) (tunnel.Endpoint, error) {
-	id := stream.ID()
-	ss.Lock()
-	abp, ok := ss.awaitingBidiPipeMap[id]
-	if ok {
-		delete(ss.awaitingBidiPipeMap, id)
-	}
-	ss.Unlock()
-
-	if !ok {
-		return nil, nil
-	}
-	dlog.Debugf(ctx, "   FWD %s, connect session %s with %s", id, abp.stream.SessionID(), stream.SessionID())
-	bidiPipe := tunnel.NewBidiPipe(abp.stream, stream)
-	bidiPipe.Start(ctx)
-
-	defer close(abp.bidiPipeCh)
-	select {
-	case <-ss.Done():
-		return nil, status.Error(codes.Canceled, "session cancelled")
-	case abp.bidiPipeCh <- bidiPipe:
-		return bidiPipe, nil
-	}
-}
-
-func (ss *sessionState) Cancel() {
-	ss.cancel()
-	close(ss.dials)
-}
-
-func (ss *sessionState) Dials() <-chan *rpc.DialRequest {
-	return ss.dials
-}
-
-func (ss *sessionState) Done() <-chan struct{} {
-	return ss.ctx.Done()
-}
-
-func (ss *sessionState) LastMarked() time.Time {
-	return ss.lastMarked
-}
-
-func (ss *sessionState) SetLastMarked(lastMarked time.Time) {
-	ss.lastMarked = lastMarked
-}
-
-func (s *State) newSessionState(now time.Time) sessionState {
-	ctx, cancel := context.WithCancel(s.ctx)
-	return sessionState{
-		ctx:        ctx,
-		cancel:     cancel,
-		lastMarked: now,
-		dials:      make(chan *rpc.DialRequest),
-	}
-}
-
-type clientSessionState struct {
-	sessionState
-	name string
-	pool *tunnel.Pool
-}
-
-type agentSessionState struct {
-	sessionState
-	agent        *rpc.AgentInfo
-	dnsRequests  chan *rpc.DNSRequest
-	dnsResponses map[string]chan *rpc.DNSResponse
-}
-
-func (ss *agentSessionState) Cancel() {
-	close(ss.dnsRequests)
-	for k, lr := range ss.dnsResponses {
-		delete(ss.dnsResponses, k)
-		close(lr)
-	}
-	ss.sessionState.Cancel()
-}
 
 // State is the total state of the Traffic Manager.  A zero State is invalid; you must call
 // NewState.
@@ -195,10 +44,10 @@ type State struct {
 	//  7. `cfgMapLocks` access must be concurrency protected
 	//  8. `cachedAgentImage` access must be concurrency protected
 	//  9. `interceptState` must be concurrency protected and updated/deleted in sync with intercepts
-	intercepts      watchable.Map[*rpc.InterceptInfo]
-	agents          watchable.Map[*rpc.AgentInfo]        // info for agent sessions
-	clients         watchable.Map[*rpc.ClientInfo]       // info for client sessions
-	sessions        map[string]SessionState              // info for all sessions
+	intercepts      watchable.Map[*rpc.InterceptInfo]    // info for intercepts, keyed by intercept id
+	agents          watchable.Map[*rpc.AgentInfo]        // info for agent sessions, keyed by session id
+	clients         watchable.Map[*rpc.ClientInfo]       // info for client sessions, keyed by session id
+	sessions        map[string]SessionState              // info for all sessions, keyed by session id
 	agentsByName    map[string]map[string]*rpc.AgentInfo // indexed copy of `agents`
 	interceptStates map[string]*interceptState
 	timedLogLevel   log.TimedLevel
@@ -422,11 +271,7 @@ func (s *State) addClient(sessionID string, client *rpc.ClientInfo, now time.Tim
 	if oldClient, hasConflict := s.clients.LoadOrStore(sessionID, client); hasConflict {
 		panic(fmt.Errorf("duplicate id %q, existing %+v, new %+v", sessionID, oldClient, client))
 	}
-	s.sessions[sessionID] = &clientSessionState{
-		sessionState: s.newSessionState(now),
-		name:         client.Name,
-		pool:         tunnel.NewPool(),
-	}
+	s.sessions[sessionID] = newClientSessionState(s.ctx, client.Name, now)
 	return sessionID
 }
 
@@ -469,13 +314,7 @@ func (s *State) AddAgent(agent *rpc.AgentInfo, now time.Time) string {
 		s.agentsByName[agent.Name] = make(map[string]*rpc.AgentInfo)
 	}
 	s.agentsByName[agent.Name][sessionID] = agent
-
-	s.sessions[sessionID] = &agentSessionState{
-		sessionState: s.newSessionState(now),
-		dnsRequests:  make(chan *rpc.DNSRequest),
-		dnsResponses: make(map[string]chan *rpc.DNSResponse),
-		agent:        agent,
-	}
+	s.sessions[sessionID] = newAgentSessionState(s.ctx, agent, now)
 
 	for interceptID, intercept := range s.intercepts.LoadAll() {
 		// Check whether each intercept needs to either (1) be moved in to a NO_AGENT state
